Move upstream tag check to sync release task
It has to be run in the task directly, and not as a pre-check,
since pre-check is part of `process_message` task that doesn't
run in the long-running worker. Long-running worker is needed here
as the Local project is being accessed.
